cosmos-nosql-vector-search/front/requirements.txt
python-dotenv
streamlit
openai
azure-cosmos

langchain_community
langchain
langchain_openai
pymongo
dnspython
tenacity
ipykernel
tiktoken